// shipeasy/src/app/models/vendor-master.ts
export interface PartyMaster {
    partymasterId: string;
    ImportExport: string;
    TDSNature: string;
    TDS_limit: string;
    TDS_utilised: string;
    accountCode: string | null;
    activeflag: boolean;
    adCode: string;
    addressInfo: {
      address: string;
      countryId: string;
      countryISOCode: string;
      countryName: string;
      stateId: string;
      stateName: string;
      cityId: string;
      cityName: string;
      postalCode: string;
      stateCode: string;
    };
    annualTurnover: string;
    bankName: string;
    branch: {
      branch_name: string;
      branch_address: string;
      barnch_country: string;
      branch_countryName: string;
      branch_state: string;
      branch_stateName: string;
      branch_cityId: string;
      branch_city: string;
      pinCode: string;
      stateCodeBranch: string;
      placeofSupply: string;
      partyCode: string;
      cust_acc_no: number;
      remarks: string;
      pic: string;
      pic_phone: number;
      pic_email: string;
      tax_name: string;
      tax_number: string;
      BROKERAGE_PAYABLE: boolean;
      creadirCustomer: boolean;
      active_flag: boolean;
      kyc_flag: boolean;
      TDS_GST_APPLICABLE: boolean;
      documents: any; // Update the type according to the data structure
    }[];
    brokerageAccCode: string;
    brokeragePayable: boolean;
    chaId: string;
    checkAcceptStatus: boolean;
    chequeAcceptance: string;
    createdBy: string | null;
    createdOn: string;
    creditCustomer: boolean;
    creditDays: string;
    currency: any; // Update the type according to the data structure
    customerAccCode: string;
    customerAccNumber: string;
    customerName: string;
    customerStatus: string;
    customerType: {
      item_id: string;
      item_text: string;
    }[];
    deactivationDate: string;
    deactivationReason: string;
    discountedBlFees: number;
    expFlag: boolean;
    expScCode: string;
    exportTeun: string;
    fasCode: string;
    faxNo: {
      faxCountryCode: string;
      faxAreaCode: string;
      faxNo: string;
    };
    flatRateDetention: string;
    groupCompany: boolean;
    gstNo: string;
    impFlag: boolean;
    impScCode: string;
    importTeun: string;
    isSez: string;
    isShippingLine: boolean;
    isTDSExempt: boolean;
    kycFlag: boolean;
    location: string;
    locationId: string;
    mloId: string;
    modifiedBy: string | null;
    module: string;
    name: string;
    operationUser: string;
    opsCode: string;
    opsFrom: string;
    opsLocation: string;
    opsName: string;
    orgId: string;
    panNo: string;
    primaryMailId: string;
    primaryNo: {
      primaryCountryCode: string;
      primaryAreaCode: string;
      primaryNo: string;
    };
    principle: string;
    proformaRequiredExp: boolean;
    proformaRequiredImp: boolean;
    rateCheckBL: boolean;
    rateCheckBooking: boolean;
    rateCheckDraftBL: boolean;
    rateCheckRO: boolean;
    refundAccCode: string;
    reg_UN_UIN: string;
    remarks: string;
    residentStatus: boolean;
    sacCode: string;
    saleCode: string;
    saleFrom: string;
    saleLocation: string;
    saleName: string;
    salePrinciple: string;
    serviceTax: string;
    serviceTaxExmp: string;
    serviceTaxExmpNo: number;
    serviceTaxLastDate: string;
    serviceTaxLastModify: string;
    serviceTaxNo: number;
    shortName: string;
    status: boolean;
    tanNo: string;
    tdsCertificate: boolean;
    tdsCertificateValid: string;
    tdsFlag: boolean;
    tdsGSTApplicable: boolean;
    tdsPercent: number;
    tenantId: string;
    tosite: string;
    updatedBy: string;
    updatedOn: string;
    url: string;
    amount: number;
    customer: any[]; // Update the type according to the data structure
    lumsumDateFrom: string;
    lumsumDateTo: string;
    isRegister: boolean;
    daysinPeriod: number;
    updatedByUID: string;
  }

  export  interface Location {
    locationName: string;
    portType: string;
    country: string;
    state: string;
    masterType: string;
    agentBranch: string;
    CFS: boolean;
    ICD: boolean;
    Yard: boolean;
    name: string;
    code: string;
    portName: string;
    terminal: string;
    EDICode: string;
    DOaddress: string;
    address: string;
    contactPerson: string;
    email: string;
    primaryCountryCode: string;
    primaryNo: string;
    DOCode: string;
    bondNo: string;
    creditDays: string;
    lineReference: boolean;
    countryISOCode: string;
    stateId: string;
    portId: string;
    status: boolean;
    module: string;
    tenantId: string;
    agentBranchId: string;
    referenceId: string;
    locationId: string;
    createdOn: string;
    updatedOn: string;
    createdBy: string;
    createdByUID: string;
    updatedBy: string;
    updatedByUID: string;
    __v: number;
  }